Translator/requirements.txt
googletrans==3.0.0
textblob==0.17.1
Pillow==9.2.0